+++ sdk/resources/azure-mgmt-resource/_metadata.json
@@ -1,5 +1,5 @@
 {
-  "commit": "97f789aeb52adfc1e20c386005839f5276874d7d",
+  "commit": "07b9f3536b0cbcdbe18a4f7037af272b1df5bbc1",
   "repository_url": "https://github.com/Azure/azure-rest-api-specs",
   "autorest": "3.10.2",
   "use": [
