+++ sdk/resources/azure-mgmt-resource/azure/mgmt/resource/databoundaries/models/_data_boundary_mgmt_client_enums.py
