+++ sdk/resources/azure-mgmt-resource/generated_tests/test_data_boundary_mgmt_data_boundaries_operations.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 import pytest
-from azure.mgmt.resource.databoundaries.v2024_08_01 import DataBoundaryMgmtClient
+from azure.mgmt.resource.databoundaries import DataBoundaryMgmtClient
 
 from devtools_testutils import AzureMgmtRecordedTestCase, RandomNameResourceGroupPreparer, recorded_by_proxy
